refactor: move shared proxy helpers to lib file
